trader/requirements.txt
requests>=2.31.0
pandas>=2.0.0
numpy>=1.24.0
ta>=0.11.0
python-dotenv>=1.0.0
websocket-client>=1.6.0
eth-account>=0.9.0
hyperliquid-python-sdk>=0.21.0
pynostr>=0.7.0